requirements.txt
atlassian-python-api==3.20.1
certifi==2021.10.8
charset-normalizer==2.0.12
Deprecated==1.2.13
docxcompose==1.3.4
idna==3.3
lxml==4.8.0
oauthlib==3.2.0
pypiwin32==223
python-docx==0.8.11
pywin32==303
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
urllib3==1.26.9
wrapt==1.14.0